pyproject.toml
[tool.poetry]
name = "contacts-fastapi"
version = "0.1.0"
description = ""
authors = ["zhendler <zhendlerbing@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.5"}
pydantic-settings = "^2.6.1"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
passlib = "^1.7.4"
bcrypt = "^4.2.0"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
